Add 2nd Compressor Temperatur Parameter 95 and 96
95: heating_temperature_outside_2nd_compressor
Setting of the minimum outside temperature from which the 2nd compressor can be released as required. Above the set outside temperature, the 2nd compressor remains blocked.

96: hotwater_temperature_forerun_2nd_compressor
Flow 2nd compressor domestic hot water Setting of the flow temperature from which domestic hot water is prepared with the second compressor to optimize the charging time and the achievable domestic hot water temperatures.